requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
numpy==1.25.2
opencv-python-headless==4.8.0.76
scikit-learn==1.3.0
Pillow==10.0.0
pytesseract==0.3.10
python-multipart==0.0.6
cairosvg==2.7.1